+++ src/main/java/bdv/tools/brightness/presetfunc/LogPresetFunc.java
@@ -35,7 +35,7 @@ public class LogPresetFunc extends AbstractPresetFunc
 {
 	private static final double K = 20.0;
 
-	public LogPresetFunc( final float min, final float max, final int paletteRangeLength )
+	public LogPresetFunc( final double min, final double max, final int paletteRangeLength )
 	{
 		super( min, max, paletteRangeLength );
 	}
@@ -47,7 +47,7 @@ public class LogPresetFunc extends AbstractPresetFunc
 	}
 
 	@Override
-	public LogPresetFunc withRange( final float min, final float max )
+	public LogPresetFunc withRange( final double min, final double max )
 	{
 		return new LogPresetFunc( min, max, getPaletteRangeLength() );
 	}
